+++ src/tracker/platform/camera_control_macosx.cpp
@@ -69,3 +69,15 @@ void camera_control_restore_system_settings(CameraControl* cc,
 {
     // Not implemented yet
 }
+
+bool
+camera_control_get_frame_layout(CameraControl *cc, int width, int height, struct CameraControlFrameLayout *layout)
+{
+    return camera_control_fallback_frame_layout(cc, width, height, layout);
+}
+
+int
+camera_control_get_preferred_camera()
+{
+    return -1;
+}